requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
beautifulsoup4==4.11.2
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.1.3
cloudstorage==0.11.0
colorama==0.4.6
distlib==0.3.6
Django==4.1.6
filelock==3.9.0
Flask==2.2.2
frozenlist==1.3.3
future==0.18.3
idna==3.3
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==2.4.1
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.1
pandas==1.5.3
platformdirs==2.6.2
pycodestyle==2.10.0
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2022.7.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.7
virtualenv==20.17.1
Werkzeug==2.2.2
yarl==1.8.2